+++ src/app/components/recruitment-boards/detail/RecruitmentBoardDetail.tsx
@@ -2,19 +2,49 @@ import { RecruitmentBoardTitle } from '@/app/components/recruitment-boards/detai
 import RecruitmentBoardContent from '@/app/components/recruitment-boards/detail/RecruitmentBoardContent';
 import dayjs from 'dayjs';
 import { RecruitmentBoardsApi } from '@/app/lib/types/recruitmentBoards/recruitmentBoards';
+import ApplyButton from '@/app/components/recruitment-boards/detail/ApplyButton';
+import ModifyButton from '@/app/components/recruitment-boards/detail/ModifyButton';
+import CheckApplicantButton from '@/app/components/recruitment-boards/detail/CheckApplicantButton';
+import styles from '@/app/recruitment-boards/detail/page.module.scss';
 
 export default function RecruitmentBoardDetail({
   boardDetail,
-  buttonBlock,
+  userId,
+  writerUserId,
+  boardId,
 }: {
   boardDetail: RecruitmentBoardsApi;
-  buttonBlock: JSX.Element;
+  userId: number;
+  writerUserId: number;
+  boardId: string;
 }) {
   const { success, data } = boardDetail;
 
   if (!data.title) {
     return <p>Loading...</p>;
   }
+
+  const buttonBlock = (
+    <div className={styles.buttonBlock}>
+      {userId === writerUserId ? (
+        <>
+          <ModifyButton />
+          <CheckApplicantButton
+            id={boardId}
+            title={boardDetail.data.title}
+            boardType={boardDetail.data.type}
+            tag={boardDetail.data.tag}
+          />
+        </>
+      ) : (
+        <ApplyButton
+          title={boardDetail.data.title}
+          detail={boardDetail.data.summary}
+          tag={boardDetail.data.tag}
+        />
+      )}
+    </div>
+  );
 
   return (
     <main>
@@ -29,10 +59,10 @@ export default function RecruitmentBoardDetail({
         target={data.recruitmentTarget}
         recruitmentNum={data.recruitmentNum}
         recruitmentStart={dayjs(data.recruitmentStart).format(
-          'YYYY.MM.DD HH:mm',
+          'YYYY.MM.DD HH:mm'
         )}
         recruitmentDeadline={dayjs(data.recruitmentDeadline).format(
-          'YYYY.MM.DD HH:mm',
+          'YYYY.MM.DD HH:mm'
         )}
         activity={data.activityCycle}
         activityStart={dayjs(data.activityStart).format('YYYY.MM.DD HH:mm')}
